Transfer of functionality to the admin panel. Changing View Routing
