+++ Workers/Workers.Web/Areas/Admin/Controllers/AdminController.cs
@@ -24,7 +24,7 @@ namespace Workers.Web.Areas.Admin
 
         [HttpGet("all")]
         //[Authorize(Policy = "ForEmail")]
-        [Authorize(Roles = "Admin, user")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> GetAll()
         {
             var roles = await _db.Roles.ToListAsync();
