+++ Workers/Workers.Web/Areas/Admin/Views/Employee/Edit.cshtml
@@ -2,6 +2,7 @@
 
 @{
     ViewData["Title"] = "Edit";
+    Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
 <h1>Edit</h1>
@@ -10,7 +11,8 @@
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form asp-action="EditEmployee">
+        <form asp-area="Admin" asp-controller="Employee" asp-action="EditEmployee" method="post"
+              >
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <div class="form-group">
                 <input asp-for="Id" type="hidden" class="form-control" />
@@ -26,10 +28,10 @@
                 <span asp-validation-for="LastName" class="text-danger"></span>
             </div>
             @*<div class="form-group">
-                <label asp-for="Login" class="control-label"></label>
-                <input asp-for="Login" class="form-control" />
-                <span asp-validation-for="Login" class="text-danger"></span>
-            </div>*@
+            <label asp-for="Login" class="control-label"></label>
+            <input asp-for="Login" class="form-control" />
+            <span asp-validation-for="Login" class="text-danger"></span>
+        </div>*@
             <div class="form-group">
                 <label asp-for="Email" class="control-label"></label>
                 <input asp-for="Email" class="form-control" />
@@ -43,7 +45,7 @@
 </div>
 
 <div>
-    <a asp-action="GetAll">Back to List</a>
+    <a asp-action="GetAll" >Back to List</a>
 </div>
 
 @section Scripts {
