+++ Workers/Workers.Web/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using Workers.Web
+@using Workers.Web.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
