+++ Workers/Workers.Web/Controllers/AccountController.cs
@@ -47,12 +47,12 @@ namespace Workers.Web.Controllers
                     _db.Employees.Add(user);
                     await _db.SaveChangesAsync();
 
+                    var role = await _db.Roles.AsNoTracking().FirstOrDefaultAsync(s => s.Name == CustomRole.Employee);
+
                     var empRole = new EmployeeRole
                     {
                         EmployeeId = user.Id,
-                        RoleId = (await _db.Roles.AsNoTracking().FirstOrDefaultAsync(x => x.Name == CustomRole.Employee)).Id
-                        //RoleId = (await _db.Roles.AsNoTracking().Select(c => new Role { Id = c.Id }).FirstOrDefaultAsync(x => x.Name == CustomRole.Employee)).Id
-
+                        RoleId = role.Id
                     };
 
                     await _db.EmployeeRoles.AddAsync(empRole);
