+++ Workers/Workers.Web/Views/Employee/GetDetails.cshtml
@@ -43,6 +43,5 @@
     </dl>
 </div>
 <div>
-    @Html.ActionLink("Edit", "Edit", new { id = Model.Id }) |
     <a asp-action="GetAll">Back to List</a>
 </div>
