+++ Workers/Workers.Web/Views/Shared/_Layout.cshtml
@@ -18,42 +18,43 @@
                 </button>
                 <div class="navbar-collapse collapse d-sm-inline-flex justify-content-between">
                     <ul class="navbar-nav flex-grow-1">
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="Employee" asp-action="GetAll">Employees</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="ProjectEmployee" asp-action="GetAll">Project Employees</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="Position" asp-action="GetAll">Position</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="Project" asp-action="GetAll">Projects</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="StatusProject" asp-action="GetAll">Status Project</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="Task" asp-action="GetAll">Tasks</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="StatusTask" asp-action="GetAll">StatusTask</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="TaskStatus" asp-action="GetAll">TaskStatus</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="EmployeesRole" asp-action="GetAll">EmployeesRole</a>
-                        </li>
-                        @*@if (User.HasClaim(x => x.Value == "email"))
+                        @if (User.Identity.IsAuthenticated)
+                        {
+                            //AdminEmployeeController
+                            <li class="nav-item">
+                                <a class="nav-link text-dark" asp-controller="Employee" asp-action="GetAll">Employees</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link text-dark" asp-controller="ProjectEmployee" asp-action="GetAll">Project Employees</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link text-dark" asp-controller="Position" asp-action="GetAll">Position</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link text-dark" asp-controller="Project" asp-action="GetAll">Projects</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link text-dark" asp-controller="StatusProject" asp-action="GetAll">Status Project</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link text-dark" asp-controller="Task" asp-action="GetAll">Tasks</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link text-dark" asp-controller="StatusTask" asp-action="GetAll">StatusTask</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link text-dark" asp-controller="TaskStatus" asp-action="GetAll">TaskStatus</a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link text-dark" asp-controller="EmployeesRole" asp-action="GetAll">EmployeesRole</a>
+                            </li>
+                            @if (User.IsInRole("Admin"))
                             {
                                 <li class="nav-item">
                                     <a class="nav-link text-dark" asp-area="Admin" asp-controller="Admin" asp-action="GetAll">Admin</a>
                                 </li>
-                            }*@
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="Admin" asp-controller="Admin" asp-action="GetAll">Admin</a>
-                        </li>
+                            }
+                        }
                         @if (!User.Identity.IsAuthenticated)
                         {
                             <li class="nav-item">
